+++ test/test_integration_cluster.rb
@@ -46,7 +46,7 @@ class TestIntegrationCluster < TestIntegration
     Process.kill :INT , @pid
     t.join
 
-    assert_match "Thread TID", output.join
+    assert_match "Thread: TID", output.join
   end
 
   def test_usr2_restart