+++ test/test_integration_single.rb
@@ -101,7 +101,7 @@ class TestIntegrationSingle < TestIntegration
     Process.kill :INT , @pid
     t.join
 
-    assert_match "Thread TID", output.join
+    assert_match "Thread: TID", output.join
   end
 
   def test_processed_requests_counter_incremented