+++ test/test_puma_server.rb
@@ -727,7 +727,9 @@ EOF
 
     sock.gets
 
-    assert_operator request_body_wait, :>=, 1000
+    # Could be 1000 but the tests get flaky. We don't care if it's extremely precise so much as that
+    # it is set to a reasonable number.
+    assert_operator request_body_wait, :>=, 900
   end
 
   def test_request_body_wait_chunked
@@ -743,6 +745,8 @@ EOF
 
     sock.gets
 
-    assert_operator request_body_wait, :>=, 1000
+    # Could be 1000 but the tests get flaky. We don't care if it's extremely precise so much as that
+    # it is set to a reasonable number.
+    assert_operator request_body_wait, :>=, 900
   end
 end