+++ test/test_pumactl.rb
@@ -1,16 +1,20 @@
 require_relative "helper"
 require_relative "helpers/config_file"
+require_relative "helpers/ssl"
 
+require 'pathname'
 require 'puma/control_cli'
 
 class TestPumaControlCli < TestConfigFileBase
+  include SSLHelper
+
   def setup
     # use a pipe to get info across thread boundary
     @wait, @ready = IO.pipe
   end
 
   def wait_booted
-    line = @wait.gets until line =~ /Listening on/
+    line = @wait.gets until line =~ /Use Ctrl-C to stop/
   end
 
   def teardown
@@ -49,7 +53,7 @@ class TestPumaControlCli < TestConfigFileBase
   end
 
   def test_environment_without_rack_env
-    with_env("RACK_ENV" => nil) do
+    with_env("RACK_ENV" => nil, 'RAILS_ENV' => nil) do
       control_cli = Puma::ControlCLI.new ["halt"]
       assert_nil control_cli.instance_variable_get("@environment")
 
@@ -91,7 +95,7 @@ class TestPumaControlCli < TestConfigFileBase
     puma_config_file = "config/puma.rb"
     development_config_file = "config/puma/development.rb"
 
-    with_env("RACK_ENV" => nil) do
+    with_env("RACK_ENV" => nil, 'RAILS_ENV' => nil) do
       with_config_file(puma_config_file, port) do
         control_cli = Puma::ControlCLI.new ["halt"]
         assert_equal puma_config_file, control_cli.instance_variable_get("@config_file")
@@ -138,18 +142,43 @@ class TestPumaControlCli < TestConfigFileBase
     assert_match "200 OK", body
     assert_match "embedded app", body
 
-    status_cmd = Puma::ControlCLI.new(opts + ["status"])
-    out, _ = capture_subprocess_io do
-      status_cmd.run
-    end
-    assert_match "Puma is started\n", out
-
-    shutdown_cmd = Puma::ControlCLI.new(opts + ["halt"])
-    out, _ = capture_subprocess_io do
-      shutdown_cmd.run
-    end
-    assert_match "Command halt sent success\n", out
+    assert_command_cli_output opts + ["status"], "Puma is started"
+    assert_command_cli_output opts + ["stop"], "Command stop sent success"
 
     assert_kind_of Thread, t.join, "server didn't stop"
+  end
+
+  def test_control_ssl
+    host = "127.0.0.1"
+    port = find_open_port
+    url = "ssl://#{host}:#{port}?#{ssl_query}"
+
+    opts = [
+      "--control-url", url,
+      "--control-token", "ctrl",
+      "--config-file", "test/config/app.rb",
+    ]
+
+    control_cli = Puma::ControlCLI.new (opts + ["start"]), @ready, @ready
+    t = Thread.new do
+      control_cli.run
+    end
+
+    wait_booted
+
+    assert_command_cli_output opts + ["status"], "Puma is started"
+    assert_command_cli_output opts + ["stop"], "Command stop sent success"
+
+    assert_kind_of Thread, t.join, "server didn't stop"
+  end
+
+  private
+
+  def assert_command_cli_output(options, expected_out)
+    cmd = Puma::ControlCLI.new(options)
+    out, _ = capture_subprocess_io do
+      cmd.run
+    end
+    assert_match expected_out, out
   end
 end